+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.33.0 (2025--)
 
 - **BREAKING CHANGES**: The `excludePattern` option has been renamed to `excludeByGlobPattern`. The old name is still available, but we plan to remove it in a future major release.
+- Add `excludeByFolderDepth` option (#213)
 - Removed references to deprecated options `excludeFiles`, `excludeFolders`, `convertSameNameSubFileToGroupIndexPage`
 
 ## 1.32.1 (2025-07-05)